refactor(react-pi): adopt createRuntimeExtras and split accessor hooks/types
replace the hand-rolled symbolPiRuntimeExtras brand + guards with the shared
createRuntimeExtras helper from @assistant-ui/core/internal; move the four accessor
hooks into runtime/hooks.ts and PiRuntimeOptions/PiRuntimeExtras into
runtime/runtimeTypes.ts (cycle-free, preserving the browser-safe boundary); type
the previously-any aui selectors; add @assistant-ui/core + @assistant-ui/store as
deps. the pure reducer/projection/host-ui, the PiThreadController +
useSyncExternalStore bridge, and the node transport are untouched, and the public
barrel is preserved exactly. unifying the two optimistic-message strategies is a
deliberate follow-up (behavior-changing).
