+++ packages/react-pi/package.json
@@ -52,6 +52,8 @@
     "test:watch": "vitest"
   },
   "dependencies": {
+    "@assistant-ui/core": "^0.2.18",
+    "@assistant-ui/store": "^0.2.18",
     "use-effect-event": "^2.0.3"
   },
   "peerDependencies": {
