+++ packages/react-pi/src/runtime/piExtras.ts
@@ -1,0 +1,5 @@
+import { createRuntimeExtras } from "@assistant-ui/core/internal";
+import type { PiRuntimeExtrasInternal } from "./runtimeTypes";
+
+export const piExtras =
+  createRuntimeExtras<PiRuntimeExtrasInternal>("usePiRuntime");
